Data Science Mark Sheet
This is a list of current marks I've achieved having completed Datacamp courses, quizzes and unit tests.  I have designed my own elaborate way of marking my own course.  In the Datacamp courses, I base my overall grade on how fast and accurate I complete each of the exercises.  The excel sheet calculator I use calculates a final grade for each Datacamp course based on a combination of how fast and accurate I complete the exercises and projects as well as the amount of hints and assistance I needed to achieve completion of each exercise.  The unit tests are based on interview questions and answers (multiple choice and written response) I obtain from credible institutions.  The written response questions are marked based on the key components provided for each answer to the question.  Some questions are worth as much as 40 marks and some are only worth 1, 2 or 3 marks.  The midterm and final exam is based on these kind of interview questions from my own library.