Current final grades achieved
This course is still ongoing since I haven't completed it yet.  It's a report card summarizing my grades with each course component.  The final exam consists of 4 components; 100 multiple choice questions, 100 marks worth of written interview questions and two different online assessments that give me a final grade as a rank out of 100.  This totals to 400 possible marks for the final exam.  Each component has it's own time allowed but the online assessments are do not have a time.  The midterm is composed of multiple choice and written response interview questions.  The course work is the final grade from the course work component (see the course work mark sheet in the main branch of this repository).